[SW-198690] Import numba packing func when needed
